Add Duration model.
The current model attempts to make arbitrarily nested tuplets
supported at a fundamental level by treating Durations like
fractions of a whole note, where the numerator may be either
an integer or a recursively defined Duration (whose value is
a fraction of the value of its outer Duration).

This model provides a 1-to-1 correspondence between the fully
reduced fractional duration and the amount of time taken by
the duration, while still allowing for non-reduced representations
for intuitive managing of tuplets.

Validation is provided against durations which would only be
possible with ties (ie, 5/8).

The basic value of the duration and dot counts are calculated
at init time.

This class is intended to be immutable.