(re-)Implement durations and metered time as just another Unit.
This commit replaces the previous in-progress implementation of
metered time, where metered time and graphical locations were
on track to be treated very differently, with a completely
new implementation that treats metered time as a subset of
graphical time.

A new Beat class merges the rational time Duration class
into a Unit subclass which can seamlessly be converted
to and from other Unit types.

The default implementation in `models/beat.py` is semi-abstract.
It does not define a _base_units_per_self_unit ratio.
The reasoning behind this is that a Beat is both a graphical and
logical construct. Beats of the same value may have different
visual lengths depending on the layout needs of the score.
It is expected, instead, that users allow the containing Staff
(or any layout system really) to create local concrete Beat classes
with concrete base unit conversion ratios.

This allows horizontal layout in a score to be controlled almost
entirely by means of layout controller objects which simply
redefine the ratio of a Beat to GraphicUnits.

The primary way users should add musical material is through
a series of `staff.add_*` methods, passing rhythmic locations
and durations in terms of tuples which represent Beat.__init__
arg signatures. The Staff passes this to the context-local
concrete Beat type in effect.

This appears to be an extremely promising solution to reconciling
metered time and graphical positioning in scores elegantly.

Problems remaining with this implementation:

  * StaffObjects now rely on explicit dependency injection of
    concrete Beat types. This is a good solution if their direct
    creation is not considered public API, but many users
    may still want to explicitly create their own objects
    without going through the staff.add_* methods,
    in which case this is awkward.
  * This currently relies on a purely proportional layout
    system with some arbitrary beat-to-staffunit ratio.
  * Currently an object's graphical position is tightly coupled
    with its logical time position. For *most* objects, this
    makes perfect sense and is desirable, but there are exceptions.
    For example, a Clef and a TimeSignature might sensibly
    occupy the same time position at the beginning of a measure,
    but this doesn't mean they should overlap visually.