Remove all parent functionality from `interface` package
This is a long-overdue completion of the move away from
using parents at the interface/qt level. Higher level objects
should explicitly position all interface objects relative to
their logical parents, rather than depending on relative position
computation in the Qt layer.

This is necessary because the complex needs of flowable coordinate
systems are not always consistent with the more simplistic Qt model.