Merge Glyph into TextObject
Upon review of similarities and differences between the two classes
it became apparent that there was really no reason to keep them
apart. The differences in implementation between the two were
arbitrary and could just as easily be applied to the other.

Hopefully this is a step toward a more unified treatment of text
across the API.