Implement alternate glyph lookups in MusicFont
This allows glyph_info() to look up metadata about
alternate glyphs by passing an optional alternate number.

This also introduces a MusicFontGlyphNotFoundError exception,
which is now raised when a glyph_info() call lookup fails
to find a glyph (instead of returning None as before).

Furthermore, this introduces a canonicalName field to the
returned glyph info dictionary.