Remove AnchoredPoint and non-0-page Points from Path
This modifies the drawing signatures of Path significantly,
preferring to let Path handle the implementation details
of translating points to PathElements.

Path drawing methods no longer take a page field in their
position descriptions. This simplifies the Path drawing API
while restricting the possibility for path elements to cross
pages. This is preferable, as a use case for such a feature
is hard to imagine.

This is a step toward moving away from and removing the
AnchoredPoint class entirely. The next class to remove it
from is likely Spanner.