Remove AnchoredPoint from Spanner and simplify its API
This removes a long-standing redundancy with Spanners,
where their position and parent is duplicated. Previously,
it was always implied that Spanner.start would have the
same position and parent as the object itself, leading
to a lot of confusion and redundancy.

This also adds the Spanner.length property, to be utilized
in future Spanner classes such as RepeatingMusicTextSpanner,
which will need to know its length.