Move music text bounding rect calculation to MusicFont
And fixes RepeatingMusicTextLine's repetition counting
by querying the MusicFont for the bounding rect width
of a single repetition of its virtual base text.

Previously this was impossible since music text bounding
rect calculations relied strictly on an existing MusicTextObject
whose `text` property was the text in question. Since in
RepeatingMusicTextLine, we need to know the size of text
which is different than the `text` property, it was necessary
to move this functionality elsewhere.