Move MusicFont bounding rect calculation back to MusicTextObject
The previous commit was in error, not taking into account the
fact that MusicChars actually have their own Font reference,
leading to potentially inconsistent behavior.

This adds a _char_list_bounding_rect() method to MusicTextObject
which takes a list of MusicChars and determines what their
bounding rect would be if they were hypothetically used
in the object.