Remove Staff.contents & surrounding logic
This has been replaced in favor of using the more general
GraphicObjects.descendants pattern, which searches not only
top-level children of the staff, but also the entire descendant
object tree.

The Contents class has been left in place for now, with the
expectation that it will prove relevant later on when higher-level
musical contexts are implemented.