Completely overhaul page system
This removes the page attribute from Point and AnchoredPoint,
and implements in its place actual Page objects, generated
on demand by the global Document via a new PageSupplier class.

To place a GraphicObject on a specific page, simply parent it
to the page in question. Instead of doing:

  SomeObject(Point(0, 0, page=2), parent=None)

Do:

  SomeObject(Point(0, 0), parent=brown.document.pages[2])

Instead of top-level GraphicObjects being parented to the
global root Document, they are now attached to Page objects.
The default parent for objects is now brown.document.pages[0]

Page acts mostly like a GraphicObject, but is not actually one.
This was necessary to avoid circular imports, but in the future
it would definitely be better to find a way to let it be an
actual GraphicObject subclass.

All GraphicObjectInterface classes now know *nothing* about pages.
It is up to the GraphicObject classes which create them to compute
their canvas positions and pass those to them. This is highly
preferable, as it helps reduce the footprint of the interface
classes further.

This new pattern vastly simplifies all logic relating to coordinates
and removes much of the complexity around pages.

Other changes caught up in this weekend tirade include:
* Addressing several TODO's
* Removing several redundant methods, especially those relating
  to coordinate mapping
* Removing some unnecessary classes:
  - NewPage (just use a NewLine with a different page parent)
  - AutoNewPage (^)
  - MockGraphicObject (InvisibleObject suffices for testing)
* Improving documentation in various classes

Happy 400th commit! 🎂 🎂 🎂