Implement PedalLine
This primitive implements the line-style pedal notation,
allowing users to pass in a list of half-lift locations
where a small triangular lift appears in the line.
Half lift positions may be either relative to the start
position or relative to other objects in the staff.

The surprising ease of implementing this fairly
complex object is telling of the progress being
made in the API.