Cut several interface defaults & improve consistency
This is part of a move toward not using default
values anywhere in the interface package, preferring
to let higher-level classes pass defaults downward
for more consistency and clarity.

This change also adds a brush property to TextInterface
and QEnhancedTextItem.

Additionally, from_existing constructors have been
added to Pen and Brush, and GraphicObject has been
refactored to have a class attribute for default pens
and brushes, which are now cloned to create default
pens and brushes in the pen and brush setters.
This allows default brushes and pens to be controlled
on a class-by-class basis - notably, allowing Pen
to have a default invisible brush, while other objects
have a solid color one.