AnchoredPoint -> ParentPoint and fix Beam
During the renaming AnchoredPoint process I accidentally
noticed that Beam was completely broken and got involved
with fixing it. This reimplements the way Beam is constructed
and adds tests to it.

This also adds a utility testing assertion for PathElement
which allows an equality assertion between arbitrary PathElement
objects comparing all of their relevant properties.