Make OctaveLine no longer automatically transpose
While writing docs I realize that this has a lot of problems,
specifically that it requires the octave line exists before chord rests
under it are placed, which is very counterintuitive. It also goes
against the general design principle where as much as possible is
graphical.

If later we're able to use a deferred chordrest construction model, we
can revisit this.