Implement a new system of staff fringes
This is a major change with several important changes included:

- Staff and TabStaff have been unified under the AbstractStaff class,
through which common behavior is shared
- A new class, StaffGroup has been added which enables inter-staff
contextual knowledge for staff system needs. This will soon to
integrated into MultiStaffObject
- Staff and TabStaff now both support fringes which occupy the left
margin of flowables. StaffFringeLayout encodes layout on a per-system
basis, and is now used to position objects in fringes
- Staff and TabStaff now both automatically furnish margin controllers
to their flowables such that the fringe fits within the margin
- KeySignature has been mostly re-implemented to be much more
efficient and to play well with the fringe layout system.
- PositionedObject now supports a new annotation
`@render_cached_property`, which allows properties to be cached only
at render time.