Improve path bounding rect pen compensation
This is a small improvement related to #14 (poor handling of thick
pens), but doesn't fix the main issue. This corrects the
bounding/clipping rect calculation to use the Qt pen's `widthF()`
instead of `width()`, allowing sub-point precision. It also removes an
incorrect scaling adjustment, since scaling applies to the painter as
a whole and automatically compensates for scaling.

Credit to @Xavman42 for shining light on this issue.