Fix pen width padding in bounding/clipping rects
The key problem was that the QClippingPath's padding was set from the
pen value before the pen was set, so it was always using the default
pen width even when the drawn path had a different pen. The solution
was to simply defer the padding calculation to occur within
`update_geometry()`.

This change also fixes a few other related issues:

- The padding derived from pen width is now `width / 2`, not just the
  width, since Qt divides the pen stroke width evenly across the ideal
  path line.
- Text paths were also failing to update their clipping geometry after
  setting the pen.
- Paths were unnecessarily calling `update_geometry` twice for every
  render. This change adds a QClippingPath init parameter
  `defer_geometry_calculation` which allows us to improve performance
  by avoiding this.

Fixes #14 thanks to help from @Xavman42