Fix viewport scaling bug
The logic for deriving the current viewport scale was incorrect,
and the logic for setting it depended on that as well. This made
viewport scaling difficult and there were some cases where it caused
divide-by-zero crashes. This fix simply resolves the issue by having
the viewport scale setter use the same crude logic used by the getter.

Fixes #89