Fix breakable_length on all built-in spanners
Spanner and Spanner2D provide a custom implementation of
PositionedObject.breakable_length, but most concrete built-in spanners
did not actually get this custom behavior due to multiple-inheritance
method resolution order issues.

This fixes the issue by reordering the relevant superclass
declarations, adds tests around this, and adds documentation warning
future users. It's not an ideal solution but it's the architecture we
have.

Fixes #71