Support missing glyphAdvanceWidths in SMuFL metadata
And add new example `other_music_fonts.py`

This closes #29 - I verified neoscore mostly supports other SMuFL
fonts but building out a new example demonstrating it and testing it
locally with the Sebastian font. I discovered a minor bug in the
process and fixed it. I've also updated the docs to describe this
capability.